dokończony logger, na ten moment jedynie serializuje kule

+++ Logika/Logger.cs
@@ -31,28 +31,60 @@ namespace Logika
         public Logger(List<SphereLogic> spheres)
         {
             Spheres = spheres;
-            startLogger();
+            startLogger(1000);
         }
         public Logger() { }
 
-        public void startLogger()
+        public void startLogger(int interval)
         {
+            string currentDirectory = Directory.GetCurrentDirectory();
+            Console.WriteLine("Tak Logger sie wykonuje..");
             enable = true;
+
+            string directoryPath = "Logs";
+            try
+            {
+                // Create the directory
+                if (Directory.Exists(directoryPath))
+                {
+                    string fileName = ".\\Logs\\log.json";
+                    stream = File.Create(fileName);
+                }
+                else
+                {
+                    Directory.CreateDirectory(directoryPath);
+                    string fileName = ".\\Logs\\log.json";
+                    stream = File.Create(fileName);
+                }
+                
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"An error occurred: {ex.Message}");
+            }
+
+
+            Console.WriteLine(currentDirectory);
+            
+        
+
+
+
             Task loggSpheres = new Task(async () =>
             {
                 while (enable)
                 {
-                    string fileName = "log.json";
-                    stream = File.Create(fileName);
+                    
 
                     /*await createStream.DisposeAsync();*/
                     watch.Start();
-                    if (watch.ElapsedMilliseconds > 1000)
+                    if (watch.ElapsedMilliseconds > interval)
                     {
                         watch.Restart();
                         foreach (SphereLogic sphereLogic in spheres)
                         {
                             await JsonSerializer.SerializeAsync(stream, sphereLogic);
+                            //Console.WriteLine("Tak Logger sie wykonuje..");
                         }
 
                     }
@@ -60,6 +92,8 @@ namespace Logika
                 }
             }
                         );
+            loggSpheres.Start();
+
         }
         public void Stop()
         {
